// src/api/gen/types/ServiceRequest.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

export type ServiceRequest = {
    /**
     * @description
     * Format: `int32`
     * @type integer
    */
    durationMinutes: number;
    /**
     * @minLength 1
     * @type string
    */
    name: string;
    price: number;
};